refactor: query client 위치 이동
