+++ src/main.tsx
@@ -3,7 +3,7 @@ import { createRoot } from "react-dom/client"
 import { QueryClientProvider } from "@tanstack/react-query"
 import "./index.css"
 import App from "./App.tsx"
-import { queryClient } from "./shared/lib/query/index.ts"
+import { queryClient } from "./shared/api/query-client.ts"
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools"
 
 createRoot(document.getElementById("root")!).render(
